+++ CHANGELOG.md
@@ -13,3 +13,9 @@ All notable changes to this project will be documented in this file.
 ## [1.2.1] - 2015-11-03 ##
 ### Updated ###
 - architecture;
+
+## [1.2.2] - 2015-11-24 ##
+### Updated ###
+- access to settings;
+- configuration generator;
+- other minor points;